#186: Add Coordinator Comments section to Opportunity profile
