+++ src/components/Dashboard/Profile/sections/Comments/common/index.ts
@@ -1,0 +1,7 @@
+export * from "./styles";
+export { Comment } from "./Comment";
+export { CommentActionMenu } from "./CommentActionMenu";
+export { DeleteCommentDialog } from "./DeleteCommentDialog";
+export { useCommentDelete } from "./hooks/useCommentDelete";
+export { useCommentEdit } from "./hooks/useCommentEdit";
+export { useCommentMenu } from "./hooks/useCommentMenu";
