+++ src/components/Dashboard/Profile/sections/Comments/index.ts
@@ -1,1 +1,3 @@
 export { Comments } from "./Comments";
+export { OpportunityComments } from "./opportunity";
+export { VolunteerComments } from "./volunteer";
