+++ src/components/Dashboard/Profile/sections/Comments/opportunity/index.ts
@@ -1,0 +1,1 @@
+export { OpportunityComments } from "./OpportunityComments";
