+++ src/components/Dashboard/Profile/sections/Comments/volunteer/index.ts
@@ -1,0 +1,1 @@
+export { VolunteerComments } from "./VolunteerComments";
