+++ src/hooks/useUpdateComment.ts
@@ -1,15 +1,18 @@
 import { apiPathComment } from "@/config/constants";
 import { useMutationQuery } from "@/hooks";
+import { Id } from "need4deed-sdk";
 
 type UpdateCommentData = {
   text: string;
 };
 
-export const useUpdateComment = (volunteerId: number, commentId: number) => {
+type EntityType = "volunteer" | "opportunity";
+
+export const useUpdateComment = (entityId: Id, commentId: number, entityType: EntityType = "volunteer") => {
   return useMutationQuery<UpdateCommentData, { message: string }>({
     apiPath: `${apiPathComment}/${commentId}`,
     method: "patch",
     successMessage: "dashboard.commentsSection.commentUpdated",
-    queryKeyToInvalidate: ["volunteer", String(volunteerId)],
+    queryKeyToInvalidate: [entityType, String(entityId)],
   });
 };
